Move NLP report prompt to prompts folder
