--- graphrag/index/operations/summarize_communities_text/__init__.py
@@ -1,4 +1,0 @@
-# Copyright (c) 2024 Microsoft Corporation.
-# Licensed under the MIT License
-
-"""Package to summarize communities using text unit text."""
